requirements.txt
langchain-core
langchain
langchain-huggingface
huggingface-hub
python-dotenv
reportlab
